+++ examples/print_joint_positions.cpp
@@ -22,7 +22,7 @@ std::ostream& operator<<(std::ostream& ostream, const std::array<T, N>& array) {
 
 int main(int argc, char** argv) {
   if (argc != 2) {
-    std::cerr << "Usage: ./print_joint_positions <robot-hostname>" << std::endl;
+    std::cerr << "Usage: " << argv[0] << " <robot-hostname>" << std::endl;
     return -1;
   }